requirements.txt
Django==1.11.17
djangorestframework==3.9.0
Pillow==5.1.0
psycopg2==2.7.6.1
